[lti] Rollback instantiation cache after we exit synthesis mode
Summary:
Synthesis mode might produce a different result than checking mode in implicit instantiation, because one has access to return hint and one does not. We should use a fresh instantiation cache for synthesis mode, because we will attempt to unify the new instantiation result with the cache during checking. The unification can cause errors for Array types, when the checking mode has access to return hint and produced a different type.

Therefore, we need to rollback the cache after we exit synthesis mode.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D40901330

fbshipit-source-id: 6459504a4ddd4fc794d69ec7a2366beeba96ff51