improve precision of Insert_type_utils.is_react_redux_file_key
Summary:
The codemod infra special cases imports from `react-redux` so that they are imported properly from their namespace. The check relied on a string match of the imported file with `"react-redux_v5.x.x.js"`. Given that we're now on later versions of react-redux, this diff changes this matching so that any library starting with `react-redux` is considered a match.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D40947017

fbshipit-source-id: c3ee08f4b9e779631f4404050316bbbc13acba2a