[lti] Reverse spread in upper bounds
Summary:
In this diff, I reverse spread in implicit instantiation upper bound with rest. For every slice or type in spread state, I convert them to a type and do a $Rest operation. It doesn't even matter if we have all the types right, we just need to get the property names right, since we are only removing properties from tout.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D40964617

fbshipit-source-id: 38d4a3f7cb189f56238ec6ce20cf387931f32331