[lti] Reimplement placeholder using the new placeholder variant of AnyT
Summary:
This diff implements the proposal of replacing a tvar based placeholder with an AnyT based one.

The central goal of the replacement is to achieve one property: we can easily distinguish a placeholder type from other types. With an any-based solution, this task is trivial.

With that property, we can easily implement a function `has_placeholders`, which deeply visit the type until FullyResolved to search for placeholders. If they exist in a type t, then t won't be considered as a valid bound during implicit instantiation, and t won't be considered as a valid result for hint decomposition.

The second part of the implementation focus on making the system aware that we have constructed a placeholder. This is achieved by creating a function `Context.mk_placeholder` that replaces `Tvar.mk_placeholder`. Similar to replaced function, this function will also log in the context that we have created a placeholder, so that we can know not to cache synthesized expression when that happens.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D40925517

fbshipit-source-id: a4518cda57fcb27001ce7c4d89bd6c77edfd2931