[lti] Consider RReactChildren reason to be fresh array
Summary:
Similar to array literals, a list of react children is also literal like. Therefore, they should be considered as fresh as well. If not, we will fail in the following case in LTI:

```
declare function Wrap({children: [mixed, mixed]}): React.Node;

(<Wrap>
  {true ? <div></div> : <div></div>}
  <div></div>
</Wrap>);
```

We will seal all expressions tvars, so we will do `[React.Element, React.Element] ~> [mixed, mixed]`, which triggers `React.Element = mixed` and failed. We have to treat it as fresh so that it only triggers a subtyping flow.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D40987949

fbshipit-source-id: 975e25cf7407f32f7fbfbe9ca7a3db5a5b50059f