[normalizer] remove meaningless Bot types when normalizing upper bounds
Summary:
When normalizing the upper bounds of a tvar (with no lower bounds), we use a heuristic that
computes the intersection of the UseT-like upper bounds types of the tvar. This is often used
by codemods like annotate-implicit-instantiations.

If an upper bound happens to be an `ReposLowerT (OpenT (_, id))` where `id` has no upper
bounds, this will lead to a `Bot NoUpper` normalized type, that will itslef short-circuit
the entire intersection to `empty`. This is typically undesirable, so instead we filter out these
results.

The root of the issue here is that when normalizing upper bounds, like we do here, tvars with
no upper-bounds correspond to `mixed`, rather than `empty` (the dual of an empty list of lower
bounds). A fix for this is a bit impractical (would require wider changes), so instead we fix for
this by detecting this specific case of `Bot` before computing the above intersection.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D40963198

fbshipit-source-id: c9ce7fa19d0d0e3c58ed7af1b3f17076f23a0206