do not annotate with `any` on validation errors in implicit instantiation codemod
Summary:
The implicit instantiation codemod would add `any`/`$FlowFixMe` when a type validation
error occurred, regardless of whether `--default-any` had been set. In other codemods
we won't add such an annotation, but instead leave the relevant code as is. In this case
this behavior is slightly different. "Leaving the code as is" corresponds to an `_` in the
type argument list. This is the use of `Ast.Expression.CallTypeArg.Implicit` in
`get_implicit_instantiation_results` instead of `Explicit flowfixme_ast`
that was effectively used before.

This diff does another simplifying refactoring too: `get_annot` can directly return a
`Ast.Expression.CallTypeArg.t` constructor, instead of going through an
`Ast.Type.annotation_or_hint` first, and then matching on the result.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D41002908

fbshipit-source-id: ea3346e94ba26188d60ff88a8a934479a9d7f77a