move package info to the new heap
Summary:
for package.json files, we store the package's info (name and main fields) in a standalone `PackageHeap`, and also store them in the main `Parsing_heap` as "untyped" files.

this diff moves the package info into the main heap, by adding a "package" kind of `parse` entry (paralleling "typed" and "untyped").

after this there are only 2 remaining callers of `SharedMem.WithCache` (and 2 of `NoCache`).

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D39852393

fbshipit-source-id: eadfa411444afb096a341493116981a3098be94c