Remove dependency on own argument/attribute for callee and instantiation hints
Summary:
Currently instantiations and callee hints depend on their attributes/argument, and if a property or argument depends on this hint, this can cause self-cycles to occur: for example, `g(function f() { return 1 })` induces a cycle between `f` and itself, because `f` depends on `g(function f() { ... })`. This diff makes it so that even synthesizable arguments or attributes aren't depended on if they themselves are what is being hinted.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41049715

fbshipit-source-id: 5e10ee8ca47b508733585251be3985859092be95