Convert implicit instantiation results to tys before running the post-inference pass
Summary:
The implicit instantiation post-pass can operate over tvars that are reachable in the implicit instantiation results. When that happens, we end up fully resolving some empty tvars to EmptyT, which causes the codemod to insert empty annotations instead of the upper bounds of the tvar.

This diff fixes that issue by converting the types to tys _BEFORE_ the implicit instantiation post-pass. This keeps the results of the post-pass and during-inference implicit instantiations completely separated.

Note that the implicit instantiation results dict is used only in codemods.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D40916530

fbshipit-source-id: a32976f454fc78a8b6691937743959331a1b073d