Allow private readonly properties to be written in constructors
Summary:
Readonly properties can be written to in the constructor, but currently we ban writing to private readonly properties anywhere. This diff fixes by threading through a write context in SetPrivatePropT just as we do in SetPropT.

Changelog: [bugfixes] Allow private readonly properties to be written to in constructors, just like regular readonly properties.

Reviewed By: SamChou19815

Differential Revision: D41130671

fbshipit-source-id: 48eb23a451fbfd2a4a9d0ea1d7b945dcb53503ce