Add annotations for _ in implicit instantiation codemod
Summary:
In the first round codemod pass we are adding a lot of `_` when the normalizer fails to put out a type. This diffs enables us to replace `_` in future codemods when we run with more permissive flags. See the added test for an example.

We accomplish this by moving the place where we record the results of the implicit instantiation to `instantiate_poly_with_targs` instead of just `instantiate_poly`. This lets us do two things:
1. Record results including _
2. Add defaults to the results

The drawback here is that we over-record results when things already received all their type arguments explicitly. However, we only record these results while running the codemod so I'm not worried about this performance hit.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41138492

fbshipit-source-id: 06725a23b97010a7c508ed7babc3af5efccfa0d0