refactor restoring packages from saved state
Summary:
moves restoring packages into the RestoreHeaps timer and makes it follow the same pattern as parsed and unparsed.

I didn't make it use the workers since there aren't that many packages and the overhead of workers on Windows is so high. perhaps we could have one set of workers do all 3 kinds of jobs though.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D41061833

fbshipit-source-id: 541d39415f49fe46666c3e3c3e1af4d33cda6d0a