[contextual] Prevent type errors during hint decomposition from being added to context without rollback
Summary:
We use speculation mechanism to rollback errors that might occur during hint decomposition in each step. However, it is not guaranteed that this mechanism can rollback all errors.

In the added regression test, it is demonstrated that errors can still leak into the context without being rolled back (what's worse, the leaked error is misplaced into another file). Because after we exit the speculation mode, we will fire all the delayed action. In most cases, we are flowing a type into an empty tvar, which is pretty harmless. However, in some cases, the flow might trigger an EvalT, and EvalT can cause all kinds of new errors.

Therefore, in addition to the speculation rollback, we need to rollback errors in the end again to ensure no errors can escape.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D41105902

fbshipit-source-id: 74d1fb585e02ef95bbb4e103c1c1bb0dc38a8f20