Fix out-of-order reads from LHS of provider assignments
Summary:
When determining ordering of a name_def that includes LHS identifiers, we need to ensure that we've checked all the components necessary to give the identifier its general type for type-at-pos purposes. If the LHS is not a provider for the identifier, or if its the only provider, then we're already good: we require that the def associated with the assignment is typechecked. However, if there are two or more providers, we need to ensure that they're all typechecked before we can typecheck any expression that contains one of those providers. Otherwise, we'll get an out-of-order read--which would occur in the name_def_tests example here.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41166794

fbshipit-source-id: 5373b2b79a21429307aa55a8e64a4e5839d4fccc