haste packages provide haste modules
Summary:
removes `ReversePackageHeap` by having haste packages provide haste modules. when we resolve a name to a haste module, we check to see if it's a package and resolve it as a package if so (before, we failed to resolve it as a module, then tried to resolve a package).

this gives us a couple of things for free:
- duplicate provider warnings. previously, the last haste package with the same name won. you also could have a module and package with the same name, even though you `import ... from 'foo'` for both.
- transactions on the package heap. since we already know how to roll back a parse and haste info, we can roll back package changes as well.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D40989019

fbshipit-source-id: 028b5049040c5579c232181231894a388f669b83