refine what is a haste package
Summary:
a "haste package" is similar to a node package in a `node_modules` folder, except it can be located anywhere, and doesn't have to live in a parent `node_modules` folder -- just as haste modules are not path sensitive.

other Haste implementations have exceptions:
- our internal Hack implementation requires packages to opt in with a `"haste_commonjs": true` field in package.json
- in Metro, packages in node_modules folders follow node's resolution, not haste's

since flow only has a single Haste resolver, we try to approximate both use cases: if a package has "haste_commonjs", it is a haste package no matter where it lives (if you're using metro, don't use this key); otherwise, packages in node_modules are not haste packages.

this is still a bit of a hack. ideally we'd have separate resolvers for the two haste implementations. 🤷‍♂️

Changelog: [misc] When using `module.system=haste`, importing packages in `node_modules` folders now obeys Node's resolution algorithm, rather than being able to be required from anywhere in the project.

Reviewed By: samwgoldman

Differential Revision: D41191759

fbshipit-source-id: 3c516d3186f95379a07c4b96bf5c950800c145ab