update revdeps in prepare_update_file
Summary:
moves updating reverse dependencies into `prepare_create_file` and `prepare_update_file`.

normally when we parse a file we don't know its revdeps yet so the `Some` case is currently not used. but we *do* know the revdeps when we load from saved state; this will be used in a followup.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D40437060

fbshipit-source-id: 084339bec9d988e05f531db0ca6e068c99f0ce99