fix adding phantom dependency on module required as a package
Summary:
in D40989019 (https://github.com/facebook/flow/commit/8c54a10cf59f85bc6c2214f33f21a6bce3b4904a) I made us add a phantom dependency on `foo` if you require `foo/bar` and it isn't a haste package.

this causes a crash when a file imports `foo` (a module) and `foo/bar` (a non-existent subpath of a package), because `foo` gets added to the resolved requires AND the phantom deps; we don't dedupe them before writing the revdeps, which fails if it already exists.

we could make it not fail on duplicates, but we don't currently need to. before the blame rev, we added nonsense phantom deps (`foo/bar` will never be a provider). so now we just don't add any phantoms in this case.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D41211191

fbshipit-source-id: f2f98037f799d6dd351195b137424f5698dc9e1c