[codemod] Annotate type arguments that are widened later
Summary:
We still have one major gap between the classic inference mode and LTI that we don't have a codemod yet. That is, tvars for type arguments can be widened pre-LTI, but they are locked down in LTI mode at the call site.

In this diff, I added a codemod that tries to address the issue. I added a flag `enable_post_inference_targ_widened_check` that can only be enabled in codemod. If this flag is turned on, we will try to unify the inference solution of targ with the Pierce solution of the targ. If it produces an error, then we _guess_ that the targ tvar got widened, and we emit an error. The same annotate-implicit-instantiation codemod will use these errors to guide additional annotations. Currently, this is hidden behind `--include-widened` flag.

This codemod can handle cases like `React.useState([])` natively without any special casing. However, it sometimes produces non-idiomatic code like `someArray.map<Result, void>(v => ...)`. This is unfortunate, but I guess we have to live with it, otherwise we will have to do it manually.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D40614283

fbshipit-source-id: 74629fc3d2b581457c40bc60f979e7a7c38fc59d