[codemods] Do not replace explicit targs in implicit instantiation codemod
Summary:
We were not checking to make sure that an annotation did not already exist before trying to insert one. This lead to weird cases where
we would replace things like Map<string, _> to Map<_, number> because only the second targ was missing. Of course, this causes the first targ
to be missing.

This diff fixes that issue by checking to see if the targ is annotated before replacing it

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D41232025

fbshipit-source-id: 5402bbd897652f6aeb8669e78cc035a513c67f1e