[enums] Casting an enum value to `empty` should error
Summary:
Specifically for casting, e.g. `(X: T)`, we special case this behaviour for enums so that you can cast to the representation type: https://flow.org/en/docs/enums/using-enums/#toc-casting-to-representation-type
This is implemented with `EnumCastT`.

The problem is that `EmptyT, EnumCastT` was short circuiting. Fix this.

Changelog: [errors] Casting an enum to `empty` will now properly error, e.g. `(E.A: empty)`

Reviewed By: samwgoldman

Differential Revision: D41205483

fbshipit-source-id: e07f7994e1ad529f3e85a7770773773a45207784