move restarting due to watchman overflows to server
Summary:
currently when Watchman misses changes and the mergebase changes, the monitor assumes the server can't handle it and exits, which takes down the server. but really, the server should decide if it can handle the problem. soon, the server *will* be able to handle it if there's a saved state.

this diff makes the monitor notify the server, and the server exits in the same case that the monitor exited before. the monitor will then spawn a new server. the net effect is the same.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D40689229

fbshipit-source-id: 97e247c3539afa5c97c00f061a950101554c7d34