reinit when we miss a mergebase change
Summary:
experimental support for reinitializing from a new saved state when there is a "large" mergebase change.

when a file watcher event tells us that we've missed changes and also changed mergebases, we used to exit; now we load a new saved state on top of the running server.

the main benefit vs exiting/restarting is that we can continue to answer IDE requests while the reinit is happening.

this sort of file watcher event can happen if the change between mergebases is so large that watchman overflows and returns `is_fresh_instance`.

this is gated by the `saved_state.allow_reinit` config.

future work:

- we currently only do this when we missed changes during a rebase, because we can't otherwise recover. another case is when we *didn't* miss changes -- we accurately tracked all of the files that changed upstream. we have a "restart if faster" heuristic that should instead become "reinit if faster".

- perhaps we could avoid reloading from the saved state we're already on, if there's no better saved state for the new mergebase. instead, we'd need to recheck all the previously checked files, just like we do when watchman restarts but the mergebase doesn't change. but this isn't necessary for correctness and could be more work than reinit-ing from a saved state we already have on disk.

Reviewed By: samwgoldman

Differential Revision: D40689509

fbshipit-source-id: f59a9003b962dee7ffc3d4c5434d4dbc2e686db4