Rework AdderT machinery to be used for all binary arith operators
Summary:
Rename `AdderT` to `ArithT` and use it for all binary arith operators. It carries an `ArithKind` which differentiates plus, rshift3, and all other operators, since these behave differently (the rshift3 behavior is only relevant later for bigints).

Some of the cases in `flow_arith` are a bit awkward to ensure that error messages mostly stay the same.

Some errors are added/changed, but I think they are mostly improvements.

Reviewed By: gkz

Differential Revision: D41138404

fbshipit-source-id: f7bc46828cc41d6834a49792a84ddb91e988c3b4