Contextual typing in pattern defaults (2/n): Populate a write location for each step in destructuring
Summary:
In this diff, I made a big change on how to deal with destructurings. Previously, we only record the leaf identifiers of destructurings as write locations, and we will recursively do destructuring in one big step. In this diff, I switched to the model where we record each step of destructuring.

This will be helpful for contextual typing of expressions that appear in the pattern default, because then it can use the recorded types as hints. It will also be slightly helpful for performance, because we can avoid repeatedly do the same destructuring work.

On the implementation side, there are few interesting notes:

1. We use a new location kind PatternLoc to record these write types.
2. For the pattern defaults, we previously find its type in two separate Select steps: (1) Do a destructuring (e.g. GetProp), (2) do a DestructureT (Default). Now they are combined into a single Select step, because there is only one write location at the patten for both.
3. Since now each pattern can be intermediate steps, we will need a way to tell whether a binding is a leaf node during destructuring. This is accompolished by checking that the write loc doesn't have kind PatternLoc.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D41488200

fbshipit-source-id: cf4ca49a5dfb15392a899e41e83253e487c0b6ae