Stronger handling of object literals, including arrays and more precise annotations
Summary:
More progress on supporting synthesizable object literals. We now support arrays whose elements are all synthesizable. In addition, we now try to make error messages more useful for recursive large objects--previously, if an object literal had an unsynthesizable member that was not a function, we would point to the overall object's declaration and ask for an annotations. In my personal debugging, with a sufficiently large object, this made it very difficult to detemine *why* something was unsynthesizable. With this change, we'll now point to both the overall object and to individual elements, functions or otherwise, with missing type information.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41108407

fbshipit-source-id: 6bad7e0691001d14bd76bc5009048ddeb8b99156