compile fuzzy-path stubs in C11
Summary:
we use `for (int i = ...)` which is a c99-ism, but we do not explicitly enable c99. since we already use c++11 for the c++ parts, let's also use c11 for the C stubs.

the release builds are built by the flowtype/flow-ci docker image which is based on centos7, which has gcc 4.8.5. 4.8 defaults to c90, but should support c11 well enough for our purposes ([docs](https://gcc.gnu.org/wiki/C11Status)).

Changelog: [internal]

Differential Revision: D41549820

fbshipit-source-id: 06005f65c68fb0a31d990e0af43a662d35e1e63e