compile fuzzy-path stubs in C11
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/8958

we use `for (int i = ...)` which is a c99-ism, but we do not explicitly enable c99. since we already use c++11 for the c++ parts, let's also use c11 for the C stubs.

the release builds are built by the flowtype/flow-ci docker image which is based on centos7, which has gcc 4.8.5. 4.8 defaults to c90, but should support c11 well enough for our purposes ([docs](https://gcc.gnu.org/wiki/C11Status)).

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D41549820

fbshipit-source-id: 8f7541ae577fe11d0407cdbcc038a7a31a5f43b8