[implicit-instantiation] Reverse ResolveUnionT
Summary:
ResolveUnionT contains a nested upper bound u. When all the union items are reversed, it will do `resolved_union ~> u`. Therefore, the reverse of `ResolveUnion(..., upper)` is just the reverse of `upper`.

It's observed in real codebases that reversing this upper bound mostly has the effect of unblocking the reversal of other upper bounds.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D41564715

fbshipit-source-id: f74b20b410de948301668a0aeabb564765f25049