Resolve no lower tvars to any in post-inference implicit instantiation analysis
Summary:
Implicit instantiation relies on any-propagation to push any through to type arguments instead of short circuiting. The tvar resolver resolves things to `empty`, which does not enjoy the same propagation machinery. This can lead to very tricky error messages. For example:

```
//flow
const x = ([]).flat()[0];
```

In this program, [] gets resolved to `Array<empty>`. The signature of `flat` is:
```
flat(depth: void | 1): Array<$Call<(<X>($ReadOnlyArray<X>) => X) & (<X>(X) => X), T>>;
```

In this signature, `T` is therefore `empty` when we call it on `[]`. This empty gets passed into the $Call. We implicitly instantiate `X` to `X'` and then process the call with `empty` as the argument. we see `empty <: $ReadOnlyArray<X>` and pass by short-circuiting, leaving `X` underconstrained. This is especially bad because this $Call is positioned in a libdef, so the error we see is:

```
Error ┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈┈ /tmp/flow/flowlib_b54ab47e4bc05383_32172/core.js:813:34

X [1] is underconstrained by $Call [2]. [underconstrained-implicit-instantiation]

        810│      *
        811│      * param depth The maximum recursion depth
        812│      */
 [2][1] 813│     flat(depth: void | 1): Array<$Call<(<X>($ReadOnlyArray<X>) => X) & (<X>(X) => X), T>>;
        814│     /**
        815│      * Returns a new array with all sub-array elements concatenated into it recursively up to the
        816│      * specified depth.
```

This signature isn't problematic in LTI because we'd emit the underconstrained error as part of speculation and move on to the next branch.

There isn't a convenient way to adjust the error message directly. The return of a function gets propagated with an unknown_use, which makes sense because downstream errors shouldn't reference the call. Moreover, even if we propagated the use_op through for this specific case, there are many more constraints that will erase it depending on the operation we use that activates the EvalT.

I wanted to get opinions on two things:
1. This seems like a reasonable way forward. A better way might be to do empty propagation, but that seems like something we don't want to block on. Any other ideas?
2. Am I missing some opportunity to improve the error message here instead of resolving to any? I'd be happy to do that instead, but I just don't see it.

I'm leaving the previous diff unpublished until this one is ready to go, since the previous diff is unnecessary otherwise.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41405796

fbshipit-source-id: b8a824aae99a4d60dd3f064767dcefb861a1bf99