When producing cycle errors, only show the minimal elements to demonstrate cycle
Summary:
Currently, when we determine that a component is an illegal SCC, we print an error that includes all elements of the SCC. This is often waaaay too much information, and includes portions of the component that aren't useful in detemining the actual cycle that is the issue. For example, imagine a component like
```
C <--> A <--> B
```
where A, B, and C are name defs, A and B depend on each other and so do A and C (see that the arrows above point in both directions). If `C` is a non-recursively resolvable name def (e.g. an unannotated function), then the minimum cycle to get from `C` back to itself only goes through `A`--we don't need to include `B`, even though it's technically part of the cycle (since there is a longer path from C back to itself through B).

In this diff, we modify the error messages for cycle errors to only include the name defs that are included in the shortest path from non-recursively resolvable components back towards themselves, which we detemine using BFS. This cleans up error messages greatly, without compromising the information they provide.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41534993

fbshipit-source-id: 8763abedb59040a2e6c7704c3a85fff4ec799374