[tools] Compute optimal per-directory enforcement
Summary:
In this diff, I implemented a tool that can find the maximum number of directories you can seal by just changing the per-directory enforcement flag.

To use the tool, you first turn on some flag in the flowconfig (e.g. the LTI flag), then you run this tool. The tool will query `flow ls` and `flow status` to find the list of files with errors, and a list of files without errors (but still checked by flow). Then it will print a list of directories to seal. It finds the optimal list, in a sense that if all files in `foo/bar` are good, it will output only `foo/bar` only instead of `foo/bar/baz.js` and `foo/bar/42.js` and `foo/bar/boz`.

We use two trie trees to represent the good and errored files. We start from the root of good files. The optimality is achieved by a BFS visit. For each node in the good files trie, we only bfs down if there is a prefix match in the errored files trie.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D41249980

fbshipit-source-id: 79d203df16bbce002a2187156533a62b7819556e