Trim very long cycle error messages
Summary:
Even with the improvements in D41534993 (https://github.com/facebook/flow/commit/0d2ebd1ce71250600e4277e2a5593fbb4929bc30), it's still possible for cycle errors to be extremely large, with in some cases hundreds of components. The main purpose of this diff is to clean this up by capping the number of components that can be displayed. This diff also makes a few quality-of-life improvements to the error messages in other places.

First, we now print a maximum of 10 components of cycles, and a maximum of 10 suggested annotation positions. This is an arbitrary point, and it's not guaranteed that all the positions necessary to resolve a cycle will be printed. However, in pathological cases there doesn't appear to really be an alternative, and in the worst case a user can presumably just annotate the positions that are displayed and respond to the updated error message that will be generated with different positions.

In addition, large recursive object literals were previously really substantial contributors to massive error messages, because we would emit an annotation location for every unannotated property, whether or not it was something that could actually be annotated (as opposed to needing to be "annotatated" with a cast). We now separate the reporting of non-function properties that need a cast to be annotated in to a separate part of the error, which we only display if the number of properties that need to be casted is 5 or less (again, arbitrary). This will help with objects that are "mostly" correctly annotated, in that it will point out stragglers, but it won't spam the user with hundreds of locations for large unannotated recursive objects which probably need to be annotated at their declarations anyways.

Reviewed By: SamChou19815

Differential Revision: D41538212

fbshipit-source-id: 89e5e269ac4d440b68609c7e1851af8f32adc0f7