Fix read order for destructured assignment defaults
Summary:
The env builder has a bug for destructing with defaults: in the expression `({x = x} = {})`, the read of `x` in the default will point to the write to `x` in the assignment, rather than to the previous definition for `x`. This is only the case in assignment: variable declarators handle this correctly, and we fix it by matching the variable declarator behavior.

A similar issue was present in array defaults, in both assignments and declarations: `([x = x] = [])` would have the same issue. We fix it here by flipping the order in the ast mapper.

Reviewed By: SamChou19815

Differential Revision: D41538826

fbshipit-source-id: 7792e775ca6eeae87728d3b7f3e5c55abaf9bdca