[new_website] Create a mode that doesn't generate inline Flow errors
Summary:
In both the old and the new website, we call the Flow binary to generate all the inline errors for Flow code blocks. This is nice, but it makes the build slow and you have to install all the ocaml toolchain and dependencies just to `yarn start` the website. (The experience is especially bad, when I try to dev and test the new website from a GitHub codespace that doesn't have any ocaml stuff pre-installed.)

This feature is not needed to dev a lot of features of the website. For example, if you just want to tweak the css of the homepage, or write some docs on eslint. Therefore, I created this flag to make my life easier in these situations. I think this mode will also be helpful in the future, say we want to enable something like netlify deploy preview of the new website, then this can enable a much lighter weight build step.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D41544613

fbshipit-source-id: 143ae856d486cfd65a72490562e6e9268e1eac09