Contextual typing in pattern defaults (5/n): Profit
Summary: With all the preparations in the previous diffs, we are finally ready to contextually type expressions in pattern defaults. In this diff, I pass the hint, which is based on the record destructuring results. to pattern default expressions. The dependency analysis is also updated to take this newly available hint to account.

Reviewed By: mvitousek

Differential Revision: D41504632

fbshipit-source-id: f179c47d4db9e3f09d17197e424c2c86b832babe