[lti] Reset all flags properly with potential exceptions considered
Summary:
I found that sometimes we are not correctly reseting synthesis mode flag back to false, which is caused by not handling exceptions and rethrow it after we reset the flag.

I still haven't figured out why we are throwing exceptions during synthesis mode that are not caught lower in the stack, but it's always better to be safe.

Since I'm already here, I also made similar changes for all other flags.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D41361086

fbshipit-source-id: c61b6bb944507be42a8bed7961377494bab7204b