[lti][tests][EZ] Add explicit targ to function call in exports
Summary:
We are no longer erroring due to the commented reason here, but instead due to types-first. Types-first will require a cast here anyways, so it doesn't matter whether we annotated the targ or not for the types-first error. Therefore, we should add the explicit targ, because it will make the current inference and the LTI mode inference produce the same errors.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D41566477

fbshipit-source-id: af6f89d5590758e52ba641f145186472f8dff1af