[lti] Rollback new errors when we are synthesizing expressions for hint decomposition
Summary:
Consider the following program:

```
declare function poly<T>((string) => void, T): T;
declare function expectString(string): string;
poly((v) => {}, expectString(3));
```

In `poly((v) => {}, expectString(3));` there is a polymorphic function, and an unannotated lambda, so it will trigger the left-to-right algorithm. `expectString(3)` is a function call, so we don't know whether it can be definitely synthesized without context, so we can't fully resolve it before the left-to-right algorithm, so it will be synthesized on-demand. During the synthesis of `expectString(3)`, unsurprisingly, a type error will occur. Due to the hint decomposition setup, it will fail the entire hint decomposition, so now we will emit `missing-local-annot` error, even though it doesn't depend on the second arg at all.

In this diff, I addressed the issue by rolling back errors when synthesizing expressions. When we produce new errors, we cannot cache the result anymore since otherwise we will never recover these rolled-back errors.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D41560332

fbshipit-source-id: af64ea25c6ae9ac02356376988e987bde5eba0cf