Support recursive function statics defined as arrows
Summary:
Currently function statics that are recursive with the base function are always cycles if the statics are defined with arrow functions (rather than `function()` functions), even if annotated. This is because a `function` is always its own name def, which can be recursively resolved, while arrow functions as statics would be stored as `ExpressionDefs`, which can't be. This diff ensures that a function static arrow is always stored in its own name_def, so that a cycle between it and the base function can be resolved by annotating the arrow.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41560782

fbshipit-source-id: c0337fbf1d4f97050ffa827eff2025d754a4c09d