Stop logging recheck stats
Summary:
We use recheck stats to determine whether to restart (soon: reinit) or recheck
the changes. These stats come from before types-first and don't really reflect
costs that we care about these days:

* It doesn't distinguish between merge and check times
* It doesn't account for skipping work due to recheck opts
* It happens too late to account for EdenFS file materialization costs

Instead, we should change how we deal with rebases in lazy mode. Currently a
file becomes "focused" just because we rebased over a changed to that file.
Instead, we should only parse these files to update the dependency graph, but
only check files which have changed since mergebase.

To account for the cost of materializing these files for parsing, we would need
an estimator that runs much earlier, and would only have access to the pre-
rebase dependency graph and list of changed files.

This diff removes the places where we log this information, but does not change
how the stats are collected or how they are used.

Reviewed By: mroch

Differential Revision: D41657676

fbshipit-source-id: 61bf179acd6f0159fa36978dad6a8ef75c29d4b3