Fix instantiate_callee for overloaded callable object definitions
Summary:
Currently in LTI calling a callable object with an overloaded callable fails with an error that the uninstantiated generic type is incompatible with the actual argument type--e.g. in the test case added in this diff, we get
```
Cannot call `c` because `T` [1] is incompatible with number [2] in the first parameter. [incompatible-call]
```
This is because callable objects with overloaded call props seem to be implemented actually as an intersection of objects, not as an object with a call prop that is an intersection. This combines poorly with the type dispatching in `Type_hint.instantiate_callee` because it only knows to look for `PolyT`s, not `ObjT`s that when called may act like `PolyT`s. This diff fixes that.

Reviewed By: SamChou19815

Differential Revision: D41688951

fbshipit-source-id: e24f9c4061e9fb6a1d60898b62cba891efde3c22